refactor: flatten & rename modules
