+++ aiengine/runtime/neurdbrt/model/mlp_clf/builder.py
@@ -1,0 +1,6 @@
+from ..base.builder import BuilderBase
+from .model import MLP
+
+
+class MLPBuilder(BuilderBase):
+    pass
