+++ aiengine/runtime/neurdbrt/model/mlp_clf/model.py
@@ -1,4 +1,3 @@
-import torch
 import torch.nn as nn
