--- aiengine/runtime/neurdbrt/models/armnet/model/__init__.py
@@ -1,2 +1,0 @@
-from .layers import *
-from .model import *
